Update HOME Page
the older version is my OWN code, and this new updates is from the W3School and I basically copied it (just retype is not the same, somehow...). And it creates an new Navigation Bar which I am currently using google's web address on for now. 

It should: change the colour as your mouse hovers over the links other than the one you are currently in.