Update Video List
Added the images of the right show and got the size right for now, and I got the table border and padding and margin. And the name of shows in bold and dark blue as text colour since want it different so stand out(maybe will change the colour later). And ....ah, the images got the link (to google for now as always....) .